Cap qualified pull file names and move file naming out of the entity builder
